Move rt::options to spawn::options
Also exports SyncActorOptions from the root of the crate as it already
happens for ActorOptions.
