+++ examples/1_hello_world.rs
@@ -1,7 +1,8 @@
 #![feature(never_type)]
 
 use heph::actor;
-use heph::rt::{self, ActorOptions, Runtime, RuntimeRef, ThreadLocal};
+use heph::rt::{self, Runtime, RuntimeRef, ThreadLocal};
+use heph::spawn::ActorOptions;
 use heph::supervisor::NoSupervisor;
 
 fn main() -> Result<(), rt::Error> {
