+++ examples/3_rpc.rs
@@ -4,7 +4,8 @@ use std::fmt;
 
 use heph::actor;
 use heph::actor_ref::{ActorRef, RpcMessage};
-use heph::rt::{self, ActorOptions, Runtime, RuntimeRef, ThreadLocal};
+use heph::rt::{self, Runtime, RuntimeRef, ThreadLocal};
+use heph::spawn::ActorOptions;
 use heph::supervisor::NoSupervisor;
 
 fn main() -> Result<(), rt::Error> {
