+++ examples/4_sync_actor.rs
@@ -1,7 +1,8 @@
 #![feature(never_type)]
 
 use heph::actor::SyncContext;
-use heph::rt::{self, Runtime, SyncActorOptions};
+use heph::rt::{self, Runtime};
+use heph::spawn::SyncActorOptions;
 use heph::supervisor::NoSupervisor;
 
 fn main() -> Result<(), rt::Error> {
