+++ src/rt/local/scheduler/inactive.rs
@@ -314,9 +314,9 @@ mod tests {
     use std::sync::atomic::{AtomicUsize, Ordering};
     use std::sync::Arc;
 
-    use crate::rt::options::Priority;
     use crate::rt::process::{Process, ProcessId, ProcessResult};
     use crate::rt::RuntimeRef;
+    use crate::spawn::options::Priority;
 
     use super::{
         Branch, Inactive, Pointer, ProcessData, LEVEL_SHIFT, N_BRANCHES, POINTER_TAG_BITS,
