+++ src/rt/local/scheduler/mod.rs
@@ -13,9 +13,9 @@ use inbox::Manager;
 use log::trace;
 
 use crate::actor::NewActor;
-use crate::rt::options::Priority;
 use crate::rt::process::{self, ActorProcess, ProcessId};
 use crate::rt::ThreadLocal;
+use crate::spawn::options::Priority;
 use crate::supervisor::Supervisor;
 
 mod inactive;
