+++ src/rt/local/scheduler/tests.rs
@@ -8,9 +8,9 @@ use std::rc::Rc;
 
 use crate::actor::{self, NewActor};
 use crate::rt::local::scheduler::{ProcessData, Scheduler};
-use crate::rt::options::Priority;
 use crate::rt::process::{Process, ProcessId, ProcessResult};
 use crate::rt::{RuntimeRef, ThreadLocal};
+use crate::spawn::options::Priority;
 use crate::supervisor::NoSupervisor;
 use crate::test::{self, init_local_actor_with_inbox, AssertUnmoved};
