+++ src/rt/process/tests.rs
@@ -12,9 +12,9 @@ use std::time::Duration;
 use mio::Token;
 
 use crate::actor::{self, Actor, NewActor};
-use crate::rt::options::Priority;
 use crate::rt::process::{ActorProcess, Process, ProcessData, ProcessId, ProcessResult};
 use crate::rt::{RuntimeRef, ThreadLocal};
+use crate::spawn::options::Priority;
 use crate::supervisor::{NoSupervisor, Supervisor, SupervisorStrategy};
 use crate::test::{self, init_local_actor_with_inbox, AssertUnmoved, TEST_PID};
