+++ src/rt/shared/scheduler/inactive.rs
@@ -392,9 +392,9 @@ mod tests {
     use std::sync::atomic::{AtomicUsize, Ordering};
     use std::sync::Arc;
 
-    use crate::rt::options::Priority;
     use crate::rt::process::{Process, ProcessId, ProcessResult};
     use crate::rt::RuntimeRef;
+    use crate::spawn::options::Priority;
 
     use super::{Branch, Inactive, Pointee, Pointer, ProcessData};
