+++ src/rt/shared/scheduler/mod.rs
@@ -12,9 +12,9 @@ use inbox::Manager;
 use log::trace;
 
 use crate::actor::NewActor;
-use crate::rt::options::Priority;
 use crate::rt::process::{self, ActorProcess, Process, ProcessId};
 use crate::rt::ThreadSafe;
+use crate::spawn::options::Priority;
 use crate::supervisor::Supervisor;
 
 mod inactive;
