+++ src/rt/shared/scheduler/runqueue.rs
@@ -101,9 +101,9 @@ mod tests {
     use std::pin::Pin;
     use std::time::Duration;
 
-    use crate::rt::options::Priority;
     use crate::rt::process::{Process, ProcessId, ProcessResult};
     use crate::rt::RuntimeRef;
+    use crate::spawn::options::Priority;
 
     use super::{Node, ProcessData, RunQueue};
