+++ src/rt/shared/waker.rs
@@ -184,11 +184,11 @@ mod tests {
 
     use mio::Poll;
 
-    use crate::rt::options::Priority;
     use crate::rt::process::{Process, ProcessData, ProcessId, ProcessResult};
     use crate::rt::shared::waker::{self, WakerData};
     use crate::rt::shared::{RuntimeInternals, Scheduler};
     use crate::rt::{RuntimeRef, Timers};
+    use crate::spawn::options::Priority;
     use crate::test;
 
     const PID1: ProcessId = ProcessId(1);
