+++ tests/functional/actor_context.rs
@@ -6,8 +6,8 @@ use std::pin::Pin;
 use std::task::Poll;
 
 use heph::actor::{self, NoMessages, RecvError};
-use heph::rt::{ActorOptions, Runtime, Signal, ThreadLocal, ThreadSafe};
-use heph::spawn::Spawn;
+use heph::rt::{Runtime, Signal, ThreadLocal, ThreadSafe};
+use heph::spawn::{ActorOptions, Spawn};
 use heph::supervisor::NoSupervisor;
 use heph::test::{init_local_actor, poll_actor};
