+++ tests/functional/tcp/server.rs
@@ -11,8 +11,8 @@ use heph::actor::messages::Terminate;
 use heph::actor::{self, Actor, NewActor};
 use heph::net::tcp::server;
 use heph::net::{TcpServer, TcpStream};
-use heph::rt::{self, ActorOptions, Signal, ThreadLocal};
-use heph::spawn::Spawn;
+use heph::rt::{self, Signal, ThreadLocal};
+use heph::spawn::{ActorOptions, Spawn};
 use heph::supervisor::{NoSupervisor, Supervisor, SupervisorStrategy};
 use heph::test::init_local_actor;
 use heph::{ActorRef, Runtime};
